renamed abis.ts as hash.ts
