+++ yarn-project/circuit-types/src/packed_arguments.ts
@@ -1,5 +1,5 @@
 import { Fr, Vector } from '@aztec/circuits.js';
-import { computeVarArgsHash } from '@aztec/circuits.js/abis';
+import { computeVarArgsHash } from '../../circuits.js/dest/hash/index.js';
 import { BufferReader, serializeToBuffer } from '@aztec/foundation/serialize';
 import { FieldsOf } from '@aztec/foundation/types';
