--- yarn-project/circuits.js/src/abis/index.ts
@@ -1,1 +1,0 @@
-export * from './abis.js';
