+++ yarn-project/circuits.js/src/contract/contract_address.ts
@@ -4,8 +4,8 @@ import { pedersenHash } from '@aztec/foundation/crypto';
 import { Fr } from '@aztec/foundation/fields';
 import { ContractInstance } from '@aztec/types/contracts';
 
-import { computeVarArgsHash } from '../abis/abis.js';
 import { GeneratorIndex } from '../constants.gen.js';
+import { computeVarArgsHash } from '../hash/hash.js';
 import { PublicKey } from '../types/public_key.js';
 
 // TODO(@spalladino): Review all generator indices in this file
