+++ yarn-project/circuits.js/src/contract/contract_tree/contract_tree.ts
@@ -1,6 +1,7 @@
 import { ContractFunctionDao, Fr, FunctionData, FunctionLeafPreimage } from '@aztec/circuits.js';
-import { hashVK } from '@aztec/circuits.js/abis';
 import { FunctionSelector, FunctionType } from '@aztec/foundation/abi';
+
+import { hashVK } from '../../hash/hash.js';
 
 /**
  * Computes the hash of a hex-encoded string representation of a verification key (vk).
