+++ yarn-project/circuits.js/src/hash/index.ts
@@ -1,0 +1,1 @@
+export * from './hash.js';
