+++ yarn-project/pxe/src/note_processor/produce_note_dao.ts
@@ -1,6 +1,6 @@
 import { L1NotePayload, TxHash } from '@aztec/circuit-types';
 import { Fr, PublicKey } from '@aztec/circuits.js';
-import { computeCommitmentNonce, siloNullifier } from '@aztec/circuits.js/abis';
+import { computeCommitmentNonce, siloNullifier } from '@aztec/circuits.js/hash';
 import { AcirSimulator } from '@aztec/simulator';
 
 import { NoteDao } from '../database/note_dao.js';
